+++ packages/system-access/test/__mocks__/@sap-ux/store.ts
@@ -1,0 +1,12 @@
+const store = jest.requireActual('@sap-ux/store');
+
+const mockedService = {
+    read: jest.fn().mockReturnValue(undefined),
+    write: jest.fn()
+};
+
+module.exports = {
+    ...store,
+    mockedService,
+    getService: jest.fn().mockResolvedValue(mockedService)
+};