+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.19.15
+
+### Patch Changes
+
+-   Updated dependencies [375ca861]
+    -   @sap-ux/ui5-config@0.19.0
+    -   @sap-ux/odata-service-writer@0.14.27
+    -   @sap-ux/ui5-application-writer@0.24.1
+
 ## 0.19.14
 
 ### Patch Changes